daemon: Configure Wireguard for local node
This commit adds support in the cilium-agent for configuring the local
Wireguard device and establishing the tunnel to remote peers. This is
implemented by watching for changes to the local and remote node
objects.

Wiregurad traffic to remote nodes is redirected by setting up a route
for the pod CIDR of the remote node. This is subject to change in a
follow-up PR.

Co-authored-by: Martynas Pumputis <m@lambda.lt>
Signed-off-by: Sebastian Wicki <sebastian@isovalent.com>
Signed-off-by: Martynas Pumputis <m@lambda.lt>