wireguard: Set rp_filter=2 of direct routing device
When a pod sends a request to a remote node and BPF masq is used, the
request is not SNAT-ed and is sent over a non-wireguard tunnel device
(in most cases over the direct routing device).

As there is the route "podCIDR via wireguardIP" on the remote node, the
request is going to be dropped on the direct routing device if the
rp_filter value of it is set to "1".

To avoid that, this commit changes the rp_filter to 2 of the direct
routing device and also of the wireguard tunnel device (for the same
reason).

In addition, we should add the remote node IP to the local allowed-ips
list, as the reply from the remote node will go over the wireguard
tunnel and its source IP should be whitelisted (in the allowed-ips
list).

Co-authored-by: Sebastian Wicki <sebastian@isovalent.com>
Signed-off-by: Sebastian Wicki <sebastian@isovalent.com>
Signed-off-by: Martynas Pumputis <m@lambda.lt>