iptables: add support for NOTRACK rules for pod-to-pod traffic
When running in KPR mode, all pod-to-pod traffic goes through connection
tracking twice: the first time traffic is tracked by netfilter and the
second one by the BPF datapath.

This is suboptimal as doing conntrack at the netfilter layer is not
strictly required for the operation of the CNI, but it has some
important performance implications.

This commit introduces a new agent option,
--install-no-conntrack-iptables-rules (enabled by default) which
installs `-j NOTRACK` Iptables rules that match all traffic from and to
the native-routing-cidr CIDR, to prevent netfilter from tracking all
pod-to-pod traffic.

This feature can be only enabled when Cilium is:

* running in direct routing
* running in full KPR mode
* not running its CNI chained on top of other CNI plugins
* not running in a managed Kubernetes environment such as AKS, EKS or GKE

There are a few reasons for these specific requirements:

* when running in tunneling mode (and with full KPR), the pod-to-pod
  traffic is encapsulated and decapsulated in the context of the BPF
  datapath program and so it is already skipping the netfilter raw table
  where connection tracking happens.
* when running in kube-proxy mode conntrack is required to perform NAT
* we cannot make assumptions on how a generic CNI is going to use
  netfilter conntrack

Signed-off-by: Gilberto Bertin <gilberto@isovalent.com>